Split LogRowView
